image_generator.py
import os
from pathlib import Path
from huggingface_hub import InferenceClient
from PIL import Image

def generate_image(prompt: str, panel_index: int, output_dir: Path) -> str:
    """
    Uses the FLUX.1-dev text-to-image model to generate an image from the prompt.
    Saves to output_dir / panel_{index}.png and returns the relative URL.
    """
    token = os.environ.get("HF_TOKEN")
    filename = f"panel_{panel_index}.png"
    filepath = output_dir / filename
    relative_url = f"/static/panels/{filename}"

    try:
        if not token or token == "your_huggingface_token_here":
            raise ValueError("Invalid or missing HF_TOKEN")
            
        # We use SDXL instead of FLUX to bypass gated model access restrictions
        client = InferenceClient(model="stabilityai/stable-diffusion-xl-base-1.0", token=token)
        image = client.text_to_image(prompt)
        image.save(str(filepath))
        return relative_url
    except Exception as e:
        print(f"Error generating image: {e}")
        # Fallback to an empty placeholder image so the UI never breaks
        placeholder = Image.new("RGB", (1024, 768), color=(128, 128, 128))
        placeholder.save(str(filepath))
        return relative_url
