requirements.txt
fastapi
uvicorn[standard]
huggingface_hub
Pillow
nltk
python-dotenv
jinja2
python-multipart
requests
